IT-group-work-final/requirements.txt
psycopg2-binary
dj-database-url
django
gunicorn
whitenoise
django-simple-captcha
Pillow